# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/categories": {
            "get": {
                "description": "Позволяет получить список всех категорий. Доступно всем пользователям.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Категории"
                ],
                "summary": "Получить все категории",
                "operationId": "get-all-categories",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Номер страницы",
                        "name": "Page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Размер страницы",
                        "name": "PageSize",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "asc",
                            "desc"
                        ],
                        "type": "string",
                        "description": "По возрастанию/по убыванию",
                        "name": "Order",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Характеристика для сортировки",
                        "name": "OrderBy",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/pagination.Paging-models_Category"
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "Позволяет создать новую категорию. Доступно только для администраторов.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Категории"
                ],
                "summary": "Создать категорию",
                "operationId": "create-category",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Category"
                        }
                    },
                    "400": {
                        "description": "Параметры новой категории некорректны",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/categories/{id}": {
            "get": {
                "description": "Позволяет получить категорию по его идентификатору. Доступно всем пользователям.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Категории"
                ],
                "summary": "Получить категорию по идентификатору",
                "operationId": "get-category",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Идентификатор категории",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Category"
                        }
                    },
                    "400": {
                        "description": "Некорректный идентификатор",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Категория с заданным идентификатором не найден",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "Позволяет удалить категорию. Доступно только для администраторов.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Категории"
                ],
                "summary": "Удалить категорию",
                "operationId": "delete-category",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Идентификатор категории",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Некорректный идентификатор",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Категория с заданным идентификатором не найдена",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "409": {
                        "description": "С удаляемой категорией есть связанные задачи",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "patch": {
                "description": "Позволяет обновить данные категории. Доступно только для администраторов.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Категории"
                ],
                "summary": "Обновить категорию",
                "operationId": "update-category",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Идентификатор категории",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Category"
                        }
                    },
                    "400": {
                        "description": "Параметры категории некорректны",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Категория с заданным идентификатором не найдена",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/categories/{id}/force": {
            "delete": {
                "description": "Позволяет принудительно удалить категорию по идентификатору вместе со всеми связаннами задачами каскадно. Доступно только для администраторов.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Категории"
                ],
                "summary": "Принудительно удалить категорию со связанными задачами",
                "operationId": "delete-force-category",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Идентификатор категории",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Некорректный идентификатор",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Категория с заданным идентификатором не найдена",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/categories/{id}/has-tasks": {
            "get": {
                "description": "Позволяет получить информацию о том, есть ли связанные задачи с категорией по ID. Доступно для всех пользователей.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Категории"
                ],
                "summary": "Есть ли у категории связанные задачи",
                "operationId": "has-tasks-category",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Идентификатор категории",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controllers.HasTasks"
                        }
                    },
                    "400": {
                        "description": "Некорректный идентификатор",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Категория с заданным идентификатором не найдена",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "controllers.HasTasks": {
            "type": "object",
            "properties": {
                "hasTasks": {
                    "type": "boolean"
                }
            }
        },
        "models.Category": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "pagination.Order": {
            "type": "string",
            "enum": [
                "asc",
                "desc"
            ],
            "x-enum-varnames": [
                "Asc",
                "Desc"
            ]
        },
        "pagination.Pagination": {
            "type": "object",
            "properties": {
                "options": {
                    "$ref": "#/definitions/pagination.PagingOptions"
                },
                "total_items": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                }
            }
        },
        "pagination.Paging-models_Category": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Category"
                    }
                },
                "pagination": {
                    "$ref": "#/definitions/pagination.Pagination"
                }
            }
        },
...
        "pagination.PagingOptions": {
            "type": "object",
            "properties": {
                "order": {
                    "$ref": "#/definitions/pagination.Order"
                },
                "order_by": {
                    "type": "string"
                },
                "page": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                }
            }
        }
    }
}